requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
dj-rest-auth==7.1.1
Django==6.0
django-allauth==65.14.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.48.0
idna==3.11
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.31.0
rsa==4.9.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
